+++ test/test.js
@@ -1,11 +1,11 @@
 /*!
- * Copyright (c) 2016-2021 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2016-2022 Digital Bazaar, Inc. All rights reserved.
  */
 'use strict';
 
 const sinon = require('sinon');
-const brPassport = require('bedrock-passport');
-global.passportStub = sinon.stub(brPassport, 'optionallyAuthenticated');
+const {passport} = require('bedrock-passport');
+global.passportStub = sinon.stub(passport, 'authenticate');
 const bedrock = require('bedrock');
 require('bedrock-account');
 require('bedrock-account-http');